SO-2022: Change type of complex map group and map priority to integer
- this is necessary to support values greater than Byte.MAX_VALUE (127)

https://snowowl.atlassian.net/browse/SO-2022

Conflicts:
	snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/index/entry/SnomedRefSetMemberIndexEntry.java
	snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/index/refset/RefSetMemberMutablePropertyUpdater.java
	snomed/com.b2international.snowowl.snomed.refset.core/src/com/b2international/snowowl/snomed/refset/clone/SnomedRefSetCloner.java

+++ snomed/com.b2international.snowowl.snomed.importer.rf2/src/com/b2international/snowowl/snomed/importer/rf2/Rf2RefSetMember.java
@@ -154,13 +154,13 @@ public class Rf2RefSetMember extends UnsupportedRf2RefSetMember {
 	}
 
 	@Override
-	public byte getMapGroup() {
-		return Byte.parseByte(record.get(6));
+	public int getMapGroup() {
+		return Integer.parseInt(record.get(6));
 	}
 
 	@Override
-	public byte getMapPriority() {
-		return Byte.parseByte(record.get(7));
+	public int getMapPriority() {
+		return Integer.parseInt(record.get(7));
 	}
 
 	@Override
